An idea how to force python ownership
The problem with findGtTask is, that CPP
is still the owner of task copies, leading to ever
growing task objects.

The idea now is to steal the ownership from CPP.
This is realized, by extending the extended wrapper object
with a forcePythonOwnership variable. If this is true,
the pointed-to object will be deleted even if the CPP object
has a parent.

Closes #261